Extract `JBSnorro.Tests.BashEscapingTests` into file
